+++ languages/de/tl_article.php
@@ -1,0 +1,25 @@
+<?php
+/**
+ * Contao Open Source CMS
+ *
+ * Copyright (c) 2015 Heimrich & Hannot GmbH
+ *
+ * @package share
+ * @author  Rico Kaltofen <r.kaltofen@heimrich-hannot.de>
+ * @license http://www.gnu.org/licences/lgpl-3.0.html LGPL
+ */
+
+$lang = &$GLOBALS['TL_LANG']['tl_article'];
+
+/**
+ * Fields
+ */
+$lang['addShare'][0] = 'Syndikationen aktivieren';
+$lang['addShare'][1] = 'Hier legen Sie fest, ob der Artikel gedruckt werden kann, als PDF heruntergeladen, oder in Sozialen Netzwerken geteilt werden kann.';
+$lang['shareModule'][0] = 'Einstellungsmodul';
+$lang['shareModule'][1] = 'Wählen Sie hier ein Modul aus, welchen die Einstellung für die Syndikation enthält.';
+
+/**
+ * Legends
+ */
+$lang['share_legend'] = 'Share-Einstellungen';