+++ languages/en/tl_article.php
@@ -1,0 +1,25 @@
+<?php
+/**
+ * Contao Open Source CMS
+ *
+ * Copyright (c) 2015 Heimrich & Hannot GmbH
+ *
+ * @package share
+ * @author  Rico Kaltofen <r.kaltofen@heimrich-hannot.de>
+ * @license http://www.gnu.org/licences/lgpl-3.0.html LGPL
+ */
+
+$lang = &$GLOBALS['TL_LANG']['tl_article'];
+
+/**
+ * Fields
+ */
+$lang['addShare'][0] = 'Add Syndication';
+$lang['addShare'][1] = 'Get options to print, pdf-print or share your article.';
+$lang['shareModule'][0] = 'Settings module';
+$lang['shareModule'][1] = 'Choose a module containing the settings for the share functionality.';
+
+/**
+ * Legends
+ */
+$lang['share_legend'] = 'Share settings';